+++ tests/test_geoseries_util.py
@@ -89,6 +89,7 @@ def test_geometry_collection_extract():
     test_geometrytypes = geoseries_util.get_geometrytypes(test_gdf.geometry)
     assert len(test_geometrytypes) == 5
     test_result_gdf = test_gdf.copy()
+    assert isinstance(test_result_gdf, gpd.GeoDataFrame)
     test_result_gdf.geometry = geoseries_util.geometry_collection_extract(
         test_result_gdf.geometry, PrimitiveType.POLYGON
     )
@@ -96,7 +97,7 @@ def test_geometry_collection_extract():
         test_result_gdf.geometry
     )
     assert len(test_result_geometrytypes) == 2
-    for index, geom in test_result_gdf.iteritems():
+    for index, geom in test_result_gdf.items():
         assert geom is not None
 
 
@@ -115,6 +116,7 @@ def test_harmonize_geometrytypes():
     test_gdf_geometrytypes = geoseries_util.get_geometrytypes(test_gdf.geometry)
     assert len(test_gdf_geometrytypes) == 2
     test_result_gdf = test_gdf.copy()
+    assert isinstance(test_result_gdf, gpd.GeoDataFrame)
     test_result_gdf.geometry = geoseries_util.harmonize_geometrytypes(
         test_result_gdf.geometry
     )
@@ -123,7 +125,7 @@ def test_harmonize_geometrytypes():
     )
     assert len(test_result_geometrytypes) == 1
     assert test_result_geometrytypes[0] == GeometryType.MULTIPOINT
-    for index, geom in test_result_gdf.geometry.iteritems():
+    for index, geom in test_result_gdf.geometry.items():
         if index in [0, 1]:
             assert geom is None
         else:
@@ -143,6 +145,7 @@ def test_harmonize_geometrytypes():
     test_gdf_geometrytypes = geoseries_util.get_geometrytypes(test_gdf.geometry)
     assert len(test_gdf_geometrytypes) == 2
     test_result_gdf = test_gdf.copy()
+    assert isinstance(test_result_gdf, gpd.GeoDataFrame)
     test_result_gdf.geometry = geoseries_util.harmonize_geometrytypes(
         test_result_gdf.geometry
     )
@@ -151,7 +154,7 @@ def test_harmonize_geometrytypes():
     )
     assert len(test_result_geometrytypes) == 1
     assert test_result_geometrytypes[0] == GeometryType.MULTILINESTRING
-    for index, geom in test_result_gdf.geometry.iteritems():
+    for index, geom in test_result_gdf.geometry.items():
         if index in [0, 1]:
             assert geom is None
         else:
@@ -172,6 +175,7 @@ def test_harmonize_geometrytypes():
     # Filter the gdf a bit to test that the indexes are retained properly in
     test_gdf = test_gdf.iloc[[1, 2, 3, 4]]  # type: ignore
     test_result_gdf = test_gdf.copy()
+    assert isinstance(test_result_gdf, gpd.GeoDataFrame)
     test_result_gdf.geometry = geoseries_util.harmonize_geometrytypes(
         test_result_gdf.geometry
     )
@@ -180,7 +184,7 @@ def test_harmonize_geometrytypes():
     )
     assert len(test_result_geometrytypes) == 1
     assert test_result_geometrytypes[0] == GeometryType.MULTIPOLYGON
-    for index, geom in test_result_gdf.geometry.iteritems():
+    for index, geom in test_result_gdf.geometry.items():
         if index in [1, 2]:
             assert geom is None
         else:
@@ -202,6 +206,7 @@ def test_harmonize_geometrytypes():
     test_geometrytypes = geoseries_util.get_geometrytypes(test_gdf.geometry)
     assert len(test_geometrytypes) == 5
     test_result_gdf = test_gdf.copy()
+    assert isinstance(test_result_gdf, gpd.GeoDataFrame)
     test_result_gdf.geometry = geoseries_util.harmonize_geometrytypes(
         test_result_gdf.geometry
     )
@@ -209,12 +214,59 @@ def test_harmonize_geometrytypes():
         test_result_gdf.geometry
     )
     assert len(test_result_geometrytypes) == 5
-    for index, geom in test_result_gdf.geometry.iteritems():
+    for index, geom in test_result_gdf.geometry.items():
         if index in [0]:
             # Only None is None, empty geometry is not changed!
             assert geom is None
         else:
             assert geom is not None
+
+
+def test_is_valid_reason(tmp_path):
+    # Test with valid data + Empty geometry
+    # -------------------------------------
+    test_gdf = gpd.GeoDataFrame(
+        geometry=[  # type: ignore
+            sh_geom.Polygon(),
+            test_helper.TestData.point,
+            test_helper.TestData.multipoint,
+            test_helper.TestData.polygon_with_island,
+            test_helper.TestData.polygon_no_islands,
+            test_helper.TestData.multipolygon,
+            test_helper.TestData.geometrycollection,
+        ]
+    )
+    result = geoseries_util.is_valid_reason(test_gdf.geometry)
+
+    assert len(result) == len(test_gdf)
+    assert result.unique() == "Valid Geometry"
+
+    # Test if indexas are retained
+    # ----------------------------
+    test_filtered_gdf = test_gdf[3:-1]
+    result = geoseries_util.is_valid_reason(test_filtered_gdf.geometry)
+
+    assert len(result) == len(test_filtered_gdf)
+    assert result.unique() == "Valid Geometry"
+    assert result.index.to_list() == test_filtered_gdf.index.to_list()
+
+    # Test with None
+    # --------------
+    test_gdf = gpd.GeoDataFrame(geometry=[None])  # type: ignore
+    result = geoseries_util.is_valid_reason(test_gdf.geometry)
+
+    # is_valid_reason returns None for None geometries
+    assert result[0] is None
+
+    # Test with invalid data
+    # ----------------------
+    # Prepare test data
+    path = test_helper.get_testfile("polygon-invalid")
+    gdf = gfo.read_file(path)
+    result = geoseries_util.is_valid_reason(gdf.geometry)
+
+    assert len(result) == len(gdf)
+    assert result[0].startswith("Ring Self-intersection")
 
 
 def test_polygons_to_lines():
@@ -253,6 +305,9 @@ def test_simplify_ext(algorithm):
     "algorithm", [SimplifyAlgorithm.RAMER_DOUGLAS_PEUCKER, SimplifyAlgorithm.LANG]
 )
 def test_simplify_topo_ext(algorithm):
+    # Skip test if simplification is not available
+    _ = pytest.importorskip("simplification")
+
     input_path = test_helper.get_testfile("polygon-parcel")
     input_gdf = gfo.read_file(input_path)
     result_geoseries = geoseries_util.simplify_topo_ext(